feat: EventResponse 추가
- 이벤트 목록/상세 조회 공용 응답
- discountRate, eventPrice, totalQuantity, soldQuantity, startAt, endAt, status 포함
- Event 엔티티로부터 생성하는 from() 정적 팩토리 메서드 추가

+++ src/main/java/com/example/commerce/domain/event/dto/EventResponse.java
@@ -9,7 +9,7 @@ public record EventResponse(
         Long id,
         Long productId,
         int discountRate,
-        int eventPrice,
+        Long eventPrice,
         int totalQuantity,
         int soldQuantity,
         LocalDateTime startAt,
